+++ apidaora/asgi/app.py
@@ -1,11 +1,19 @@
 import asyncio
+from http import HTTPStatus
 from logging import getLogger
 from typing import Any, Awaitable, Callable, Dict
 from urllib import parse
 
-from ..exceptions import MethodNotFoundError, PathNotFoundError
+import orjson
+
+from ..exceptions import (
+    BadRequestError,
+    MethodNotFoundError,
+    PathNotFoundError,
+)
 from .base import ASGIApp, Receiver, Scope, Sender
 from .responses import (
+    make_json_response,
     send_method_not_allowed_response,
     send_not_found,
     send_response,
@@ -32,6 +40,34 @@ def asgi_app(router: Callable[[str, str], ResolvedRoute]) -> ASGIApp:
         else:
             route = resolved.route
 
+            if route.has_headers:
+                headers = scope['headers']
+            else:
+                headers = []
+
+            try:
+                if (
+                    hasattr(route.controller, 'middlewares')
+                    and route.controller.middlewares
+                ):
+                    for (
+                        middleware
+                    ) in route.controller.middlewares.post_routing:
+                        middleware(route.path_pattern, resolved.path_args)
+
+            except BadRequestError as error:
+                body = orjson.dumps(error.dict)
+                await send_response(
+                    send,
+                    response=make_json_response(
+                        content_length=len(body),
+                        status=HTTPStatus.BAD_REQUEST,
+                        headers=headers,
+                    ),
+                    body=body,
+                )
+                return
+
             if route.has_query:
                 query_dict = _get_query_dict(scope)
             else:
@@ -41,11 +77,6 @@ def asgi_app(router: Callable[[str, str], ResolvedRoute]) -> ASGIApp:
                 body = await _read_body(receive)
             else:
                 body = b''
-
-            if route.has_headers:
-                headers = scope['headers']
-            else:
-                headers = []
 
             response_and_body = route.controller(
                 resolved.path_args, query_dict, headers, body